[core] move AttributeBinding to gfx namespace
