+++ src/mbgl/gl/attribute.cpp
@@ -7,10 +7,10 @@ namespace gl {
 
 using namespace platform;
 
-optional<AttributeBinding> offsetAttributeBinding(const optional<AttributeBinding>& binding, std::size_t vertexOffset) {
+optional<gfx::AttributeBinding> offsetAttributeBinding(const optional<gfx::AttributeBinding>& binding, std::size_t vertexOffset) {
     assert(vertexOffset <= std::numeric_limits<uint32_t>::max());
     if (binding) {
-        AttributeBinding result = *binding;
+        gfx::AttributeBinding result = *binding;
         result.vertexOffset = static_cast<uint32_t>(vertexOffset);
         return result;
     } else {
